requirements.txt
langchain
python-dotenv
openai
ipykernel
langchain-groq
langchain-openai
langchain-core
fastapi
uvicorn
langserve
sse_starlette
streamlit
langchain-community
transformers
langchain-chroma
langchain-huggingface
sentence-transformers
bs4
langchain-classic